# docs/openapi.yaml
openapi: 3.0.3
info:
  title: iDME 工业 MiniApp API
  description: |
    华为 iDME 数据建模引擎 - 工业 MiniApp 接口文档。

    ## 快速开始

    ### 推荐创建顺序（依赖链）

    **Level 0（无依赖）**：Unit（计量单位）, BusinessPartner（往来单位）, EquipmentClassfication（设备分类）, PartClassfication（物料分类）

    **Level 1**：Department（部门）, Employee（员工）

    **Level 2**：EquipmentModel（设备型号）, Location（位置）

    **Level 3**：Equipment（设备）, Part（物料）

    **Level 4**：WorkingProcedure（工序）, WorkingPlan（工艺路线）

    **Level 5**：BOMItem（BOM项）, PlanProcedureLink（计划-工序关联）, ProcedureEquipmentLink（工序-设备关联）, ProcedurePartLink（工序-物料关联）, EquipmentSparePartLink（设备-备件关联）

    ### 注意事项
    - **引用字段**（如 `manufacturerName`, `category`）需要传入对应模块的 **ID**，不是文本名称
    - **Part** 和 **WorkingPlan** 是版本对象，`WorkingPlan.productPart` 引用 Part 时使用 `master.id` 而非实例 `id`
    - **枚举值**使用英文（如 `Standby`, `Active`, `Machinery`）
    - 所有接口的 example 字段已填充真实可用的示例ID，可直接复制使用

    ### 示例数据
    - 固定ID文档: `/docs/openapi-examples-ids.md`
    - 测试文件: `/src/test/java/com/tsukilc/idme/integration/OpenApiExamplesDataTest.java`

    ---

    - 管理界面: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services/index.html
    - 健康检查: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services/v1/health

    认证: Basic Auth 用户名 sysadmin / 密码见项目说明。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services
    description: iDME SDK 服务
security:
  - basicAuth: []
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 用户名 sysadmin，密码见项目文档
  parameters:
    pageNum:
      name: pageNum
      in: query
      description: 页码，从 1 开始
      schema:
        type: integer
        default: 1
    pageSize:
      name: pageSize
      in: query
      description: 每页大小
      schema:
        type: integer
        default: 20
    idPath:
      name: id
      in: path
      required: true
      description: 主键 ID
      schema:
        type: string
  schemas:
    ApiResponse:
      type: object
      properties:
        result:
          type: string
          example: SUCCESS
        data: {}
        errors:
          type: array
          items:
            type: string
      description: 统一响应包装
...
    EquipmentListItem:
      type: object
      description: 设备列表项
      properties:
        id: { type: string }
        equipmentCode: { type: string }
        equipmentName: { type: string }
        status: { type: string }
        locationText: { type: string }
    PartListItem:
      type: object
      description: 物料列表项
      properties:
        id: { type: string }
        partNumber: { type: string }
        partName: { type: string }
        businessVersion: { type: string }
        stockQty: { type: number }
        unit: { type: string }
...
    PartnerContact:
      type: object
      properties:
        id: { type: string }
        contactName: { type: string }
        mobile: { type: string }
        phone: { type: string }
        email: { type: string }
        role: { type: string }
        remarks: { type: string }
...
    Part:
      type: object
      properties:
        id: { type: string }
        masterId: { type: string }
        partNumber: { type: string }
        partName: { type: string }
        modelSpec: { type: string }
        stockQty: { type: integer, description: "库存数量（整数，SDK要求int32）" }
        unit: { type: string }
        supplierName: { type: string }
        category: { type: string }
        businessVersion: { type: string }
        versionNumber: { type: string }
        description: { type: string }
        drawingUrl: { type: string }
        extra: { type: object }
        createTime: { type: string, format: date-time }
        lastModifiedTime: { type: string, format: date-time }
    PartCreate:
      type: object
      required: [partNumber, partName, unit, supplierName, category]
      properties:
        partNumber:
          type: string
          description: 物料编号（建议唯一）
          example: "PART-DEMO"
        partName:
          type: string
          description: 物料名称
          example: "中心轮零件"
        modelSpec:
          type: string
          description: 型号规格
          example: "M-100"
        stockQty:
          type: integer
          description: 库存数量（整数，SDK要求int32）
          example: 100
        unit:
          type: string
          description: 计量单位ID（-> Unit模块）
          example: "1000000001"
        supplierName:
          type: string
          description: 供应商ID（-> BusinessPartner模块，SDK必填）
          example: "2000000002"
        category:
          type: string
          description: 物料分类ID（-> PartClassfication模块）
          example: "3000000002"
        businessVersion:
          type: string
          description: 业务版本号（版本对象必填）
          example: "1.0"
        description:
          type: string
          description: 描述
          example: "中心轮零件，用于传动系统"
        drawingUrl:
          type: string
          description: 图纸URL
        extra:
          type: object
          description: 扩展字段
...
    PartnerContactCreate:
      type: object
      required: [partner, contactName]
      properties:
        partner: { type: string }
        contactName: { type: string }
        mobile: { type: string }
        phone: { type: string }
        email: { type: string }
        role: { type: string }
        remarks: { type: string }
...
    Location:
      type: object
      properties:
        id: { type: string }
        locationCode: { type: string }
        locationName: { type: string }
        locationType: { type: string, enum: [Plant, Workshop, Line, Station, Warehouse] }
        parentLocation: { type: string }
        addressText: { type: string }
        manager: { type: string }
        remarks: { type: string }
        children: { type: array, items: { $ref: '#/components/schemas/Location' } }
    LocationCreate:
      type: object
      required: [locationCode, locationName, locationType]
      properties:
        locationCode:
          type: string
          description: 位置编码
          example: "LOC-DEMO"
        locationName:
          type: string
          description: 位置名称
          example: "一车间"
        locationType:
          type: string
          enum: [Plant, Workshop, Warehouse, Office]
          description: "位置类型：工厂/车间/仓库/办公室"
          example: "Plant"
        parentLocation:
          type: string
          description: 父位置ID（-> Location模块，可选，用于创建层级结构）
        addressText:
          type: string
          description: 地址
        manager:
          type: string
          description: 位置管理员ID（-> Employee模块，SDK必填）
          example: "4000000002"
        remarks:
          type: string
          description: 备注
...
paths:
...
  /api/part:
    get:
      tags: [物料管理]
      summary: 分页查询物料列表
      operationId: listPart
      parameters:
        - $ref: '#/components/parameters/pageNum'
        - $ref: '#/components/parameters/pageSize'
        - name: partNumber
          in: query
          schema: { type: string }
        - name: partName
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          total: { type: integer }
                          pageNum: { type: integer }
                          pageSize: { type: integer }
                          records:
                            type: array
                            items: { $ref: '#/components/schemas/PartListItem' }
    post:
      tags: [物料管理]
      summary: 创建物料
      operationId: createPart
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Part'
  /api/part/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [物料管理]
      summary: 查询物料详情
      operationId: getPart
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Part'
    put:
      tags: [物料管理]
      summary: 更新物料
      operationId: updatePart
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [物料管理]
      summary: 删除物料
      operationId: deletePart
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/part/{masterId}/checkout:
    post:
      tags: [物料管理]
      summary: 检出物料（创建工作副本）
      operationId: checkoutPart
      parameters:
        - name: masterId
          in: path
          required: true
          schema: { type: string }
          description: 物料主对象ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Part'
  /api/part/{masterId}/checkin:
    post:
      tags: [物料管理]
      summary: 检入物料（保存为新版本）
      operationId: checkinPart
      parameters:
        - name: masterId
          in: path
          required: true
          schema: { type: string }
          description: 物料主对象ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Part'
...
  /api/partner-contact:
    post:
      tags: [联系人]
      summary: 创建联系人
      operationId: createPartnerContact
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartnerContactCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/partner-contact/by-partner/{partnerId}:
    get:
      tags: [联系人]
      summary: 查询往来单位的所有联系人
      operationId: getPartnerContactsByPartner
      parameters:
        - name: partnerId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/PartnerContact' }
  /api/partner-contact/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    put:
      tags: [联系人]
      summary: 更新联系人
      operationId: updatePartnerContact
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartnerContactCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [联系人]
      summary: 删除联系人
      operationId: deletePartnerContact
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/location:
    post:
      tags: [位置]
      summary: 创建位置
      operationId: createLocation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocationCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/location/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [位置]
      summary: 查询位置详情
      operationId: getLocation
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Location'
    put:
      tags: [位置]
      summary: 更新位置
      operationId: updateLocation
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocationCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [位置]
      summary: 删除位置
      operationId: deleteLocation
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/location/{id}/equipments:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [位置]
      summary: 查询某位置下的所有设备
      operationId: getLocationEquipments
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/EquipmentListItem' }
...